add user drawer list component
